+++ pywren_ibm_cloud/job/job.py
@@ -1,7 +1,6 @@
 import os
 import sys
 import time
-import json
 import textwrap
 import pickle
 import logging